+++ test/flutter_simple_dependency_injection_test.dart
@@ -45,29 +45,29 @@ class ObjectB {
 }
 
 void main() {
-  test('non-genric map test', () async {
-    final injector = Injector.getInjector();
+  test('non-generic map test', () async {
+    final injector = Injector();
     injector.map((i) => ObjectC('C', 'B'));
     final instance = injector.get<ObjectC>();
     expect(instance != null, true);
   });
 
   test('can get default injector instance', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     expect(injector != null, true);
   });
 
   test('can get named injector instance', () async {
-    final injectorDefault = Injector.getInjector();
-    final injectorNamed = Injector.getInjector('name');
+    final injectorDefault = Injector();
+    final injectorNamed = Injector('name');
     expect(injectorDefault != null, true);
     expect(injectorNamed != null, true);
     expect(injectorNamed != injectorDefault, true);
   });
 
   test('can have isolated injector instances', () async {
-    final injectorOne = Injector.getInjector('one');
-    final injectorTwo = Injector.getInjector('two');
+    final injectorOne = Injector('one');
+    final injectorTwo = Injector('two');
     injectorOne
         .map<ObjectWithNoDependencies>((i) => ObjectWithNoDependencies());
     final instance = injectorOne.get<ObjectWithNoDependencies>();
@@ -115,7 +115,7 @@ void main() {
   });
 
   test('can map generic types', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithGenerics<String>>(
         (injector) => ObjectWithGenerics('Hello'));
     injector.map<ObjectWithGenerics<int>>((injector) => ObjectWithGenerics(10));
@@ -134,7 +134,7 @@ void main() {
   });
 
   test('can map class factory and get object instance', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithNoDependencies>(
         (injector) => ObjectWithNoDependencies());
     final instance = injector.get<ObjectWithNoDependencies>();
@@ -146,7 +146,7 @@ void main() {
   test(
       'can map class factory for singleton and always get back the same object',
       () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithNoDependencies>(
         (injector) => ObjectWithNoDependencies(),
         isSingleton: true);
@@ -160,7 +160,7 @@ void main() {
   });
 
   test('can map class factory and get object simple instance', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithNoDependencies>(
         (injector) => ObjectWithNoDependencies());
     injector.map<ObjectWithOneDependency>((injector) =>
@@ -173,7 +173,7 @@ void main() {
   });
 
   test('can map same object with different keys and get instances', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithNoDependencies>(
         (injector) => ObjectWithNoDependencies());
     injector.map<ObjectWithNoDependencies>(
@@ -188,7 +188,7 @@ void main() {
   });
 
   test('can map simple named string type', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<String>((injector) => 'Jon', key: 'MyName');
     final instanceOne = injector.get<String>(key: 'MyName');
     final instanceTwo = injector.get<String>(key: 'MyName');
@@ -199,7 +199,7 @@ void main() {
   });
 
   test('can construct type with additional give parameters', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.mapWithParams<ObjectWithSomeConstructorArgDependencies>(
         (injector, p) => ObjectWithSomeConstructorArgDependencies(p['id']));
     final instanceOne = injector.get<ObjectWithSomeConstructorArgDependencies>(
@@ -210,7 +210,7 @@ void main() {
   });
 
   test('can get all instances of type', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithSomeConstructorArgDependencies>(
         (injector) => ObjectWithSomeConstructorArgDependencies('0'));
     injector.map<ObjectWithNoDependencies>(
@@ -231,7 +231,7 @@ void main() {
   });
 
   test('exception thrown when type is not known', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithOneDependency>((injector) =>
         ObjectWithOneDependency(injector.get<ObjectWithNoDependencies>()));
     expect(
@@ -244,7 +244,7 @@ void main() {
   });
 
   test('exception thrown when keyed type is not known', () async {
-    final injector = Injector.getInjector();
+    final injector = Injector();
     injector.map<ObjectWithOneDependency>((injector) => ObjectWithOneDependency(
         injector.get<ObjectWithNoDependencies>(key: 'Key')));
     expect(
@@ -253,6 +253,49 @@ void main() {
             e is InjectorException &&
             e.message ==
                 "Cannot find object factory for 'ObjectWithNoDependencies::Key'")));
+    injector.dispose();
+  });
+
+  test('can remove mapping and map again', () async {
+    final injector = Injector();
+    injector.map<ObjectWithNoDependencies>(
+        (injector) => ObjectWithNoDependencies());
+
+    expect(injector.isMapped<ObjectWithNoDependencies>(), true);
+    injector.removeMapping<ObjectWithNoDependencies>();
+    expect(injector.isMapped<ObjectWithNoDependencies>(), false);
+
+    injector.map<ObjectWithNoDependencies>(
+        (injector) => ObjectWithNoDependencies());
+    expect(injector.isMapped<ObjectWithNoDependencies>(), true);
+
+    expect(injector.get<ObjectWithNoDependencies>() is ObjectWithNoDependencies,
+        true);
+
+    injector.dispose();
+  });
+
+  test('can remove all mappings of type', () async {
+    final injector = Injector();
+    injector.map<ObjectWithNoDependencies>(
+        (injector) => ObjectWithNoDependencies());
+    injector.map<ObjectWithNoDependencies>(
+      (injector) => ObjectWithNoDependencies(),
+      key: 'one',
+    );
+    injector.map<ObjectWithNoDependencies>(
+      (injector) => ObjectWithNoDependencies(),
+      key: 'two',
+    );
+
+    expect(injector.isMapped<ObjectWithNoDependencies>(), true);
+    expect(injector.isMapped<ObjectWithNoDependencies>(key: 'one'), true);
+    expect(injector.isMapped<ObjectWithNoDependencies>(key: 'two'), true);
+    injector.removeAllMappings<ObjectWithNoDependencies>();
+    expect(injector.isMapped<ObjectWithNoDependencies>(), false);
+    expect(injector.isMapped<ObjectWithNoDependencies>(key: 'one'), false);
+    expect(injector.isMapped<ObjectWithNoDependencies>(key: 'two'), false);
+
     injector.dispose();
   });
 }